Add install script, restructure README, clean up root directory
- Add interactive install.sh with arch detection, MCP client config,
  and RAG dependency handling (macOS Intel/ARM, Linux)
- Restructure README from 600 to 167 lines; move detailed content
  (system prompt, tools reference, data structure, migration) to
  docs/GUIDE.md
- Move test files from root to tests/, docs to docs/
- Remove obsolete files: requirements.txt, mypy.ini, .roo/
- Consolidate mypy config into pyproject.toml with full strict options
- Clean up .gitignore: remove *.json glob, add orphaned data dirs
- Remove PYTHONPATH from .mcp.json (uv handles it)
- Fix pre-existing test failures: ImprovisationLevel enum comparisons,
  campaign.json fixture naming, split format default assertion
