--- mypy.ini
@@ -1,21 +1,0 @@
-[mypy]
-python_version = 3.12
-warn_return_any = True
-warn_unused_configs = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-check_untyped_defs = True
-disallow_untyped_decorators = True
-no_implicit_optional = True
-warn_redundant_casts = True
-warn_unused_ignores = True
-warn_no_return = True
-warn_unreachable = True
-strict_equality = True
-
-# Per-module options
-[mypy-mcp.*]
-ignore_missing_imports = True
-
-[mypy-pydantic.*]
-ignore_missing_imports = True
