--- requirements.txt
@@ -1,35 +1,0 @@
-annotated-types==0.7.0
-anyio==4.9.0
-authlib==1.6.0
-certifi==2025.6.15
-cffi==1.17.1 ; platform_python_implementation != 'PyPy'
-click==8.2.1
-colorama==0.4.6 ; sys_platform == 'win32'
-cryptography==45.0.4
-exceptiongroup==1.3.0
-fastmcp==2.9.2
-h11==0.16.0
-httpcore==1.0.9
-httpx==0.28.1
-httpx-sse==0.4.1
-idna==3.10
-markdown-it-py==3.0.0
-mcp==1.9.4
-mdurl==0.1.2
-openapi-pydantic==0.5.1
-pycparser==2.22 ; platform_python_implementation != 'PyPy'
-pydantic==2.11.7
-pydantic-core==2.33.2
-pydantic-settings==2.10.1
-pygments==2.19.2
-python-dotenv==1.1.1
-python-multipart==0.0.20
-rich==14.0.0
-shellingham==1.5.4
-sniffio==1.3.1
-sse-starlette==2.3.6
-starlette==0.47.1
-typer==0.16.0
-typing-extensions==4.14.0
-typing-inspection==0.4.1
-uvicorn==0.35.0 ; sys_platform != 'emscripten'
